refactor(simple_webrtc): move SourceTrack and SinkTrack out of simple_webrtc
