+++ extensions/warp-blink-wrtc/src/lib.rs
@@ -17,6 +17,7 @@
 //! - the struct implementing Blink will keep a Arc<Mutex<simple_webrtc::Controller>>, allowing calls to be initiated by the UI
 //! - the webrtc task also has that simple_webrtc::Controller
 
+mod audio;
 mod host_media;
 mod signaling;
 mod simple_webrtc;
