+++ extensions/warp-blink-wrtc/src/simple_webrtc/mod.rs
@@ -44,7 +44,6 @@ use webrtc::track::track_local::track_local_static_rtp::TrackLocalStaticRTP;
 use webrtc::track::track_remote::TrackRemote;
 
 // public exports
-pub mod audio;
 pub mod events;
 
 pub use webrtc::rtp_transceiver::rtp_codec::RTCRtpCodecCapability;
